Transparent change to JonesVector
We still initialize with Ex and Ey, because we set the basis to xhat and yhat in __init__.  But now, the field is kept internally as E1 and E2, with basis vectors b1 and b2.  Before, they were kept as Ex and Ey, even if the basis vectors could be changed.
However, for convenience, we still need Ex and Ey.  Therefore, we return Ex and Ey as computed property (simply a projection of b1 and b2 onto xhat or yHat).  To set Ex or Ey however, is only possible when the basis b1 or b2 are either xhat or yhat.  If that is not the case, then we cannot set either of them.

All tests pass...!